notifications/consumers.py
import json
from channels.generic.websocket import AsyncWebsocketConsumer
from channels.layers import get_channel_layer
from django.core.cache import cache
import jwt # PyJWT

class NotificationConsumer(AsyncWebsocketConsumer):
    """
    Real-time notification push consumer.
    Auth: JWT passed as ?token=<jwt> in query string.
    """
    async def connect(self):
        # 1. Extract token from query string
        query_string = self.scope.get("query_string", "")
        params = dict([x.split("=") for x in query_string.split("&") if "=" in x])
        token = params.get("token")

        if not token:
            await self.close(code=4001) # Unauthorized
            return

        try:
            # 2. Validate JWT
            # In production, SECRET_KEY and algorithm would be in settings
            payload = jwt.decode(token, "secret-key", algorithms=["HS256"])
            self.user_id = payload["user_id"]
        except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
            await self.close(code=4001)
            return

        # 3. Group association
        self.group_name = f"notifications_{self.user_id}"
        await self.channel_layer.group_add(self.group_name, self.channel_name)

        # Store active session in Redis for monitoring (optional but good for "Strong Hire")
        cache.set(f"ws:channel:{self.user_id}", self.channel_name, timeout=3600)

        await self.accept()

    async def disconnect(self, close_code):
        # Clean up group association and Redis session
        if hasattr(self, 'group_name'):
            await self.channel_layer.group_discard(self.group_name, self.channel_name)
            cache.delete(f"ws:channel:{self.user_id}")

    async def notification_new(self, event):
        """
        Handler for 'notification.new' events pushed from Celery.
        """
        # event contains: { "type": "notification.new", "unread_count": X, "notification": { ... } }
        await self.send(text_data=json.dumps(event))
