requirements.txt
django==5.0
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
channels==4.0.0
channels-redis==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
daphne==4.1.0
django-celery-beat==2.3.0
